+++ CESystem/MovementSystem.h
@@ -16,7 +16,7 @@ public:
 	void runSystem(std::vector<Entity*> entityList);
 
 	sf::Vector2f countScale(int rotation);
-	sf::Vector2f countVelocity(sf::Vector2f scale, float speed);
+	sf::Vector2f countVelocity(sf::Vector2f scale, int speed);
 	sf::Vector2i newPosition(sf::Vector2i position, sf::Vector2f velocity);
 
 	MovementSystem();
